+++ rust/cubestore/cubestore/src/queryplanner/serialized_plan.rs
@@ -1083,7 +1083,7 @@ impl SerializedPlan {
         &self.schema_snapshot.index_snapshots
     }
 
-    pub fn files_to_download(&self) -> Vec<(IdRow<Partition>, String, Option<u64>)> {
+    pub fn files_to_download(&self) -> Vec<(IdRow<Partition>, String, Option<u64>, Option<u64>)> {
         self.list_files_to_download(|id| {
             self.partition_ids_to_execute
                 .binary_search_by_key(&id, |(id, _)| *id)
@@ -1092,7 +1092,7 @@ impl SerializedPlan {
     }
 
     /// Note: avoid during normal execution, workers must filter the partitions they execute.
-    pub fn all_required_files(&self) -> Vec<(IdRow<Partition>, String, Option<u64>)> {
+    pub fn all_required_files(&self) -> Vec<(IdRow<Partition>, String, Option<u64>, Option<u64>)> {
         self.list_files_to_download(|_| true)
     }
 
@@ -1103,6 +1103,7 @@ impl SerializedPlan {
         IdRow<Partition>,
         /* file_name */ String,
         /* size */ Option<u64>,
+        /* chunk_id */ Option<u64>,
     )> {
         let indexes = self.index_snapshots();
 
@@ -1122,6 +1123,7 @@ impl SerializedPlan {
                         partition.partition.clone(),
                         file,
                         partition.partition.get_row().file_size(),
+                        None,
                     ));
                 }
 
@@ -1131,6 +1133,7 @@ impl SerializedPlan {
                             partition.partition.clone(),
                             chunk.get_row().get_full_name(chunk.get_id()),
                             chunk.get_row().file_size(),
+                            Some(chunk.get_id()),
                         ))
                     }
                 }